Can add books in frontend w images
Implemented swagger, basic frontend html and css to be able to add the books to the server with images. Now it saves in the program. Need to find way to delete the images when ending the program and need to check if price calculation is correct and add User states next (react maybe?). Also delete book and order book states.